Add value matching functionality for the a dereferenced pointer
[Issue/Feature Description]
When trying to check the value passed via a pointer as an input to a
mock, it is possible to lose the value to stack clobbering upon
completion of the function under test.

[Resolution]
Add a new functionality to the test framework to allow for With matching
of the value behind a pointer.